+++ .gitignore
@@ -13,3 +13,5 @@ node_modules
 .DS_Store
 .idea
 .vscode
+
+.env
