+++ routes/index.js
@@ -1,30 +1,11 @@
 const router = require('express').Router();
-const { celebrate, Joi } = require('celebrate');
 const auth = require('../middlewares/auth');
 const NotFoundError = require('../errors/not-found-err');
 const { createUser, login, signout } = require('../controllers/users');
+const { signinValidation, signupValidation } = require('../validation/userValidation');
 
-router.post(
-  '/signin',
-  celebrate({
-    body: Joi.object().keys({
-      email: Joi.string().required().email(),
-      password: Joi.string().required(),
-    }),
-  }),
-  login,
-);
-router.post(
-  '/signup',
-  celebrate({
-    body: Joi.object().keys({
-      name: Joi.string().required().min(2).max(30),
-      email: Joi.string().required().email(),
-      password: Joi.string().required(),
-    }),
-  }),
-  createUser,
-);
+router.post('/signin', signinValidation, login);
+router.post('/signup', signupValidation, createUser);
 router.post('/signout', signout);
 
 router.use(auth);
